Compute G for liquid and Widom's insertion method
TI_near_crytical.py is a script that, given a number of densities compute the pressures and chemical potentials along an isotherm choosen.
widom_benchmark.py is a script that for a fixed simulation uses the widom insertion method for different number of particles.
The two postprocessing script are for obtaining the right plots ftom ehte data (in data folder).